TesseractDocker/requirements.txt
Pillow==8.3.2
pytesseract==0.3.8
Flask==2.1.1

